Rename EffectFunc to RequestDecorateFunc
