+++ src/core/frameworks/gateway/fetchAgent.ts
@@ -1,4 +1,4 @@
-import { Agent, RequestBuilder } from '@/core/frameworks/gateway/http';
+import { Agent, Request, RequestBuilder, RequestDecorateFunc } from './http';
 import { Context } from './http/context';
 
 const baseUrl = 'https://apas-todo-api.azurewebsites.net';
@@ -9,6 +9,12 @@ export class FetchAgent implements Agent {
 
   setAccessToken(token: string): void {
     this.ctx = this.ctx.set(accessTokenKey, token);
+  }
+
+  decorateAccessToken(): RequestDecorateFunc {
+    return (req: Request, ctx: Context) => {
+      req.headers['Authorization'] = `Bearer ${ctx.get(accessTokenKey)}`;
+    };
   }
 
   run(builder: RequestBuilder): Promise<any> {
