--- src/core/frameworks/gateway/http/effectFuncs.ts
@@ -1,8 +1,0 @@
-import { EffectFunc } from './requestBuilder';
-import { Request } from './request';
-
-export const setBody = (body: any): EffectFunc => {
-  return (req: Request) => {
-    req.body = body;
-  };
-};
