+++ src/core/frameworks/gateway/http/index.ts
@@ -1,5 +1,4 @@
 export * from './agent';
-export * from './effectFuncs';
+export * from './decorators';
 export * from './request';
 export * from './requestBuilder';
-export * from './token';
