+++ src/core/frameworks/gateway/http/request.ts
@@ -1,3 +1,4 @@
+export type AllowMethods = 'GET' | 'POST' | 'PUT' | 'DELETE';
 export interface Request {
   path: string;
   method: string;
